requirements.txt
absl-py==0.7.1
apturl==0.5.2
asn1crypto==0.24.0
astor==0.8.0
astroid==2.2.5
attrs==19.1.0
Automat==0.7.0
Brlapi==0.6.6
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
command-not-found==0.3
constantly==15.1.0
cryptography==2.6.1
cssselect==1.0.3
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
dj-database-url==0.5.0
Django==2.2.1
django-heroku==0.3.1
gast==0.2.2
grpcio==1.21.1
gunicorn==19.9.0
h5py==2.9.0
httplib2==0.12.3
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.20
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.1
louis==3.5.0
lxml==4.3.3
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.1.1
MarkupSafe==1.0
mccabe==0.6.1
mock==3.0.5
netifaces==0.10.4
numpy==1.16.3
oauth==1.0.1
olefile==0.45.1
opencv-contrib-python==4.1.0.25
opencv-python-headless==4.1.0.25
parsel==1.5.1
pexpect==4.2.1
Pillow==6.0.0
pipenv==2018.11.26
protobuf==3.7.1
psycopg2==2.8.2
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycups==1.9.73
PyDispatcher==2.0.5
pygobject==3.26.1
PyHamcrest==1.9.0
pylint==2.3.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==19.0.0
pypi==2.1
pyRFC3339==1.0
python-apt==1.6.4
python-debian==0.1.32
python-instagram==1.3.2
pytz==2019.1
pyxdg==0.25
PyYAML==3.12
queuelib==1.5.0
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
Scrapy==1.6.0
SecretStorage==2.3.1
selenium==3.141.0
service-identity==18.1.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
system-service==0.3
systemd-python==234
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
Twisted==19.2.0
typed-ast==1.4.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.3
usb-creator==0.3.3
vboxapi==1.0
virtualenv==16.6.0
virtualenv-clone==0.5.3
w3lib==1.20.0
wadllib==1.3.2
Werkzeug==0.15.4
whitenoise==4.1.2
wrapt==1.11.1
xkit==0.0.0
zope.interface==4.6.0
